convert auction analyticsMiddleware to ts, jest
