+++ src/typings/sharify.d.ts
@@ -61,6 +61,7 @@ declare module "sharify" {
       ARTIST_PAGE_CTA_ENABLED: string
       ARTSY_EDITORIAL_CHANNEL: string
       ARTSY_XAPP_TOKEN: string
+      AUCTION?: any
       CSRF_TOKEN: string
       CHANNEL?: { slug?: string } // Articles
       FAIR?: any // mobile fair app data
